Changed the shown messages in SmsLikeView.
Clicking on a conversation partner will now show messages with the following attributes:
A message needs to be in the current folder and needs to be sent by the selected conversation partner or it needs to be a message sent to the conversation partner and be sent from the current account.

So one comes to the conclusion that a folder hierarchy of an mail account doesnt make sense in a sms like view. In a sms like view you want to see messages from inbox and outbox to have a conversation-like view.